+++ frontend/src/components/modules/activateAccount/activateAccount.tsx
